modified cart methods

+++ FoodDeliveryApp/src/main/java/com/hexaware/hotpot/entities/Discount.java
@@ -35,9 +35,11 @@ public class Discount {
 	@JoinColumn(name = "RestaurantID")
 	private Restaurants restaurant;
 
-	@NotNull(message = "Discount percentage is required")
-	@Column(name = "DiscountPercentage")
-	private BigDecimal discountPercentage;
+
+    @NotNull(message = "Discount percentage is required")
+    @Column(name = "DiscountPercentage")
+    private int discountPercentage;
+
 
 	@NotNull(message = "Start date is required")
 	@Column(name = "StartDate")
@@ -51,7 +53,7 @@ public class Discount {
 		super();
 	}
 
-	public Discount(int discountId, Restaurants restaurant, BigDecimal discountPercentage, LocalDate startDate,
+	public Discount(int discountId, Restaurants restaurant, int discountPercentage, LocalDate startDate,
 			LocalDate endDate) {
 		super();
 		this.discountId = discountId;
@@ -77,11 +79,11 @@ public class Discount {
 		this.restaurant = restaurant;
 	}
 
-	public BigDecimal getDiscountPercentage() {
+	public int getDiscountPercentage() {
 		return discountPercentage;
 	}
 
-	public void setDiscountPercentage(BigDecimal discountPercentage) {
+	public void setDiscountPercentage(int discountPercentage) {
 		this.discountPercentage = discountPercentage;
 	}
